bleh

+++ QLPKNK_App/QLPKNK_App/DTO/ThuocDTO.cs
@@ -10,6 +10,6 @@ namespace QLPKNK_App.DTO
     {
         public string id { get; set; }
         public string name { get; set; }
-        public float price { get; set; }
+        public Double price { get; set; }
     }
 }
